requirements.txt
langchain 
langchain-community 
langchain-core 
streamlit 
python-dotenv
langchain-groq
langchain-ollama
